requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==4.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
cssselect2==0.4.1
defusedxml==0.7.1
Django==3.2.5
django-parler==2.2
dnspython==1.16.0
eventlet==0.31.1
greenlet==1.1.0
html5lib==1.1
idna==3.2
isodate==0.6.0
kombu==4.6.11
lxml==4.6.3
Pillow==8.3.1
polib==1.1.1
psycopg2-binary==2.9.1
pycparser==2.20
pyphen==0.11.0
pytz==2021.1
requests==2.26.0
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
tinycss2==1.1.0
urllib3==1.26.6
vine==1.3.0
WeasyPrint==52.5
webencodings==0.5.1
zeep==4.0.0
